Fix budget amount field size

+++ decidim-budgets/lib/decidim/api/mutations/project_attributes.rb
@@ -7,7 +7,7 @@ module Decidim
       description "Attributes for a project"
 
       argument :address, GraphQL::Types::String, "Address of this project", required: false
-      argument :budget_amount, GraphQL::Types::Int, "The budget amount for this project", required: false
+      argument :budget_amount, GraphQL::Types::BigInt, "The budget amount for this project", required: false
       argument :description, GraphQL::Types::JSON, description: "The project description", required: false
       argument :latitude, GraphQL::Types::Float, "Latitude of this project", required: false
       argument :longitude, GraphQL::Types::Float, "Longitude of this project", required: false
